refactor: 先享卡回调优化
